+++ .app/_data/app.js
@@ -1,49 +1,7 @@
-const fs = require("fs");
-
 module.exports = function () {
-  const configPath = "./../app.json";
-  const custom = fs.existsSync(configPath)
-    ? JSON.parse(fs.readFileSync(configPath))
-    : {};
-
-  return {
-    title: "Notes",
-    description: "Notes app",
-    ...custom,
-    theme: {
-      color: "sky",
-      ...custom.theme,
-    },
-    customProperties: {
-      properties: [],
-      ...custom.customProperties,
-    },
-    sidebar: {
-      links: [],
-      notes: [{}],
-      ...custom.sidebar,
-    },
-    panel: {
-      tableOfContents: true,
-      customProperties: true,
-      tags: true,
-      incomingLinks: true,
-      outgoingLinks: true,
-      externalLinks: true,
-      ...custom.panel,
-    },
-    wikilinks: {
-      autoLabel: "ref",
-      anchorLabel: "none",
-      ...custom.wikilinks,
-    },
-    tags: {
-      map: {},
-      ...custom.tags,
-    },
-    notes: {
-      pathPrefix: "/n",
-      ...custom.notes,
-    },
-  };
+  try {
+    return require("./../../app.js");
+  } catch (e) {
+    return require("./../app-config.js").defineConfig({});
+  }
 };